Done with offline heap -> Figure out how to bind to pipeline

+++ DXR-Project/Application/Timestamp.h
@@ -73,19 +73,19 @@ public:
 	FORCEINLINE static Timestamp Seconds(Float64 Seconds)
 	{
 		constexpr Float64 SECOND = 1000.0 * 1000.0 * 1000.0;
-		return Timestamp(Float64(Seconds * SECOND));
+		return Timestamp(static_cast<Uint64>(Seconds * SECOND));
 	}
 
 	FORCEINLINE static Timestamp MilliSeconds(Float64 Milliseconds)
 	{
 		constexpr Float64 MILLISECOND = 1000.0 * 1000.0;
-		return Timestamp(Float64(Milliseconds * MILLISECOND));
+		return Timestamp(static_cast<Uint64>(Milliseconds * MILLISECOND));
 	}
 
 	FORCEINLINE static Timestamp MicroSeconds(Float64 Microseconds)
 	{
 		constexpr Float64 MICROSECOND = 1000.0;
-		return Timestamp(Float64(Microseconds * MICROSECOND));
+		return Timestamp(static_cast<Uint64>(Microseconds * MICROSECOND));
 	}
 
 	FORCEINLINE static Timestamp NanoSeconds(Uint64 Nanoseconds)
